Split cloud account repository tests
